adding a watch mode action
