+++ packages/scheduler-types/src/types/TargetScheduler.ts
@@ -5,4 +5,5 @@ export interface TargetScheduler {
   abort(): void;
   run(root: string, targetGraph: TargetGraph): Promise<SchedulerRunSummary>;
   onTargetChange?(targetId: string): void;
+  cleanup(): void;
 }
